rag_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.0
python-dotenv==1.0.0
requests==2.31.0
gitpython==3.1.40
python-multipart==0.0.6
pydantic==2.5.2
faiss-cpu==1.7.4
sentence-transformers==2.2.2 